qa-visualiser/pyproject.toml
[project]
name = "qa-visualiser"
version = "0.1.0"
description = "Tool for creating questionnaire flow visualisations for Plan Tech"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.20.3",
    "pydantic>=2.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.ruff.lint]
extend-select = ["I"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
]
